pyproject.toml
[project]
name = "meta-prompter"
version = "0.1.0"
description = "A tool to generate meta-prompts from documentation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]

[project.scripts]
app = "meta_prompter:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
